Fix settings translations

+++ translations/handleTranslations.sh
@@ -60,7 +60,9 @@ do
   # delete removed l10n files that are used for language detection (they will be recreated during the write)
   find core/l10n -type f -delete
   find lib/l10n -type f -delete
-  find settings/l10n -type f -delete
+  if [ "$version" != "master" ]; then
+    find settings/l10n -type f -delete
+  fi
 
   # build JS/JSON based on translations
   /translationtool.phar convert-po-files
@@ -70,7 +72,12 @@ do
   git checkout -- tests/
 
   # create git commit and push it
-  git add apps core lib settings
+  if [ "$version" != "master" ]; then
+    git add apps core lib settings
+  else
+    git add apps core lib
+  fi
+
   git commit -am "[tx-robot] updated from transifex" || true
   git push origin $version
